pyproject.toml
[tool.poetry]
name = "usagovjobs"
version = "0.1.0"
description = "Search gov data jobs and store the results in a local db for downstream use (report email)."
authors = ["chipembere <chitiyobrian@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
SQLAlchemy = "^1.4.35"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
mypy = "^0.942"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
